backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-openai
langchain-community
chromadb
sentence-transformers
pypdf
faiss-cpu
python-dotenv
langgraph
groq
pytest
httpx
pytest-asyncio
pymongo
pymupdf
